app/requirements.txt
streamlit
python-dotenv
agno
requests
fastapi
uvicorn[standard]
python-multipart
pydantic
httpx
requests
google-genai
sqlalchemy
pymongo
pypdf
chromadb